REFACTOR/nested service call

+++ apps/api/src/email-verification/email-verification.service.ts
@@ -5,34 +5,10 @@ import { HTTPException } from "hono/http-exception";
 import { emailVerificationEmailBody } from "./email-verification.emails";
 import { sessionRepo } from "~/session/session.repo";
 import { generateSessionToken } from "~/session/session.utils";
+import { sendEmail } from "~/libs/email";
 import type { Db } from "~/libs/db";
 import type { EmailVerificationCode, User } from "~/db/db.schemas";
-import { sendEmail, type Email } from "~/libs/email";
-
-const verifyCode = async (
-  userId: EmailVerificationCode["userId"],
-  code: EmailVerificationCode["code"],
-  db: Db,
-) => {
-  const emailVerificatonCode = await emailVerificationRepo.selectByUserId(
-    userId,
-    db,
-  );
-
-  if (emailVerificatonCode?.code !== code) {
-    throw new HTTPException(401, { message: "invalid code" });
-  }
-
-  await emailVerificationRepo.deleteById(emailVerificatonCode.id, db);
-
-  const codeExpired = Date.now() >= emailVerificatonCode.expiresAt.getTime();
-
-  if (codeExpired) {
-    throw new HTTPException(401, { message: "code expired" });
-  }
-
-  await userRepo.updateEmailVerifiedAt(userId, db);
-};
+import type { Email } from "~/libs/email";
 
 const create = async (
   user: Pick<User, "id" | "email">,
@@ -65,7 +41,24 @@ const confirm = async (
   db: Db,
 ) => {
   return db.transaction(async (tx) => {
-    await emailVerificationService.verifyCode(userId, code, tx);
+    const emailVerificatonCode = await emailVerificationRepo.selectByUserId(
+      userId,
+      db,
+    );
+
+    if (emailVerificatonCode?.code !== code) {
+      throw new HTTPException(401, { message: "invalid code" });
+    }
+
+    await emailVerificationRepo.deleteById(emailVerificatonCode.id, db);
+
+    const codeExpired = Date.now() >= emailVerificatonCode.expiresAt.getTime();
+
+    if (codeExpired) {
+      throw new HTTPException(401, { message: "code expired" });
+    }
+
+    await userRepo.updateEmailVerifiedAt(userId, db);
 
     await sessionRepo.deleteByUserId(userId, tx);
 
@@ -83,5 +76,4 @@ const confirm = async (
 export const emailVerificationService = {
   create,
   confirm,
-  verifyCode,
 };
